Theme switch #9
Added a like function with animation and a switch to switch theme.
The switch is not yet connected to the rest of the page so that will be the next step.
The switch is made using the principle progressive enhancement and accessibility because it can be used with keyboard navigation.